+++ OngekiFumenEditor/Modules/SelectionFilter/ViewModels/SelectionFilterViewModel.cs
@@ -1,36 +1,31 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
-using System.Text.RegularExpressions;
 using System.Windows.Data;
+using AngleSharp.Common;
 using Caliburn.Micro;
-using NAudio.CoreAudioApi;
 using OngekiFumenEditor.Base;
 using OngekiFumenEditor.Base.EditorObjects;
 using OngekiFumenEditor.Base.EditorObjects.LaneCurve;
 using OngekiFumenEditor.Base.EditorObjects.Svg;
 using OngekiFumenEditor.Base.OngekiObjects;
 using OngekiFumenEditor.Base.OngekiObjects.Beam;
-using OngekiFumenEditor.Base.OngekiObjects.ConnectableObject;
 using OngekiFumenEditor.Base.OngekiObjects.Lane;
 using OngekiFumenEditor.Base.OngekiObjects.Wall;
-using OngekiFumenEditor.Modules.FumenEditorSelectingObjectViewer.Views;
 using OngekiFumenEditor.Modules.FumenVisualEditor.Kernel;
 using OngekiFumenEditor.Modules.FumenVisualEditor.ViewModels;
 using OngekiFumenEditor.Utils;
-using Xv2CoreLib.ACB;
 using Binding = System.Windows.Data.Binding;
 
-namespace OngekiFumenEditor.Modules.FumenEditorSelectingObjectViewer.ViewModels;
+namespace OngekiFumenEditor.Modules.SelectionFilter.ViewModels;
 
 public class SelectionFilterViewModel : ViewAware
 {
-    public FumenVisualEditorViewModel Editor { get; private set; }
+    public FumenVisualEditorViewModel Editor { get; }
 
-    public FilterOptions FilterOptions { get; set; }
+    public SelectionFilterOptionsViewModel FilterOptions { get; }
 
     public ObservableCollection<ISelectableObject> ObjectTypeFilterMatches { get; } = new();
     public ObservableCollection<ISelectableObject> OptionFilterRemovals { get; } = new();
@@ -190,207 +185,6 @@ public class SelectionFilterViewModel : ViewAware
     }
 }
 
-public enum FilterMode
-{
-    Remove,
-    Replace
-}
-
-public sealed class FilterOptions : PropertyChangedBase
-{
-    private readonly SelectionFilterViewModel SelectionFilter;
-
-    public event PropertyChangedEventHandler OptionChanged;
-
-    private FilterMode _filterSelectionChangeMode = FilterMode.Replace;
-    private YesNoFilteringMode _isCriticalFilterMode = YesNoFilteringMode.Any;
-    private LinePointTypeFilteringMode _nodeTypeFilteringMode = LinePointTypeFilteringMode.Any;
-    private LinePointTypeFilteringMode _holdPartFilteringMode = LinePointTypeFilteringMode.Any;
-    private string _tagFilter = string.Empty;
-    private bool _isTagFilterRegex = false;
-    private bool _onlyMatchLanesWithSelectedStartNodes = false;
-    private bool _onlyMatchHoldsEndsWithSelectedStartNodes = false;
-    private LeftRightFilteringMode _flickDirectionFilteringMode = LeftRightFilteringMode.Any;
-
-    public ObservableCollection<FilterBulletPalettesItem> FilterBulletPalettesItems { get; } = new();
-
-    public ObservableCollection<FilterDockableLaneOption> FilterDockableLaneOptions { get; } = new()
-    {
-        new() { Name = "WallLeft", Filter = d => d.ReferenceLaneStart.LaneType == LaneType.WallLeft},
-        new() { Name = "LaneLeft", Filter = d => d.ReferenceLaneStart.LaneType == LaneType.Left},
-        new() { Name = "LaneCenter", Filter = d => d.ReferenceLaneStart.LaneType == LaneType.Center},
-        new() { Name = "LaneRight", Filter = d => d.ReferenceLaneStart.LaneType == LaneType.Right},
-        new() { Name = "WallRight", Filter = d => d.ReferenceLaneStart.LaneType == LaneType.WallRight},
-        new() { Name = "Undocked", Filter = d => d.ReferenceLaneStart == null},
-    };
-
-    public FilterMode FilterSelectionChangeMode
-    {
-        get => _filterSelectionChangeMode;
-        set => Set(ref _filterSelectionChangeMode, value);
-    }
-
-    public YesNoFilteringMode IsCriticalFilterMode
-    {
-        get => _isCriticalFilterMode;
-        set => Set(ref _isCriticalFilterMode, value);
-    }
-
-    public LinePointTypeFilteringMode NodeTypeFilteringMode
-    {
-        get => _nodeTypeFilteringMode;
-        set => Set(ref _nodeTypeFilteringMode, value);
-    }
-
-    public LinePointTypeFilteringMode HoldPartFilteringMode
-    {
-        get => _holdPartFilteringMode;
-        set => Set(ref _holdPartFilteringMode, value);
-    }
-
-    public string TagFilter
-    {
-        get => _tagFilter;
-        set => Set(ref _tagFilter, value);
-    }
-
-    public bool IsTagFilterRegex
-    {
-        get => _isTagFilterRegex;
-        set => Set(ref _isTagFilterRegex, value);
-    }
-
-
-    public bool OnlyMatchLanesWithSelectedStartNodes
-    {
-        get => _onlyMatchLanesWithSelectedStartNodes;
-        set => Set(ref _onlyMatchLanesWithSelectedStartNodes, value);
-    }
-
-    public bool OnlyMatchHoldsEndsWithSelectedStartNodes
-    {
-        get => _onlyMatchHoldsEndsWithSelectedStartNodes;
-        set => Set(ref _onlyMatchHoldsEndsWithSelectedStartNodes, value);
-    }
-
-    public LeftRightFilteringMode FlickDirectionFilteringMode
-    {
-        get => _flickDirectionFilteringMode;
-        set => Set(ref _flickDirectionFilteringMode, value);
-    }
-
-    public FilterOptions(SelectionFilterViewModel selectionFilter)
-    {
-        SelectionFilter = selectionFilter;
-
-        // Call setters on settings to update UI
-        FilterSelectionChangeMode = _filterSelectionChangeMode;
-        IsCriticalFilterMode = _isCriticalFilterMode;
-        NodeTypeFilteringMode = _nodeTypeFilteringMode;
-        HoldPartFilteringMode = _holdPartFilteringMode;
-        TagFilter = _tagFilter;
-        IsTagFilterRegex = _isTagFilterRegex;
-        FlickDirectionFilteringMode = _flickDirectionFilteringMode;
-
-        FilterBulletPalettesItems.AddRange(selectionFilter.Editor.Fumen.BulletPalleteList.Select(bpl => new FilterBulletPalettesItem(bpl)));
-        FilterBulletPalettesItems.ForEach(i => i.PropertyChanged += (sender, args) =>
-        {
-            OptionChanged?.Invoke(sender, args);
-        });
-
-        PropertyChanged += (_, args) =>
-        {
-            OptionChanged?.Invoke(this, args);
-        };
-
-        FilterDockableLaneOptions.ForEach(x =>
-        {
-            x.PropertyChanged += (_, args) =>
-            {
-                if (args.PropertyName == nameof(FilterDockableLaneOption.IsSelected)) {
-                    OptionChanged?.Invoke(this, new PropertyChangedEventArgs(""));
-                }
-            };
-        });
-
-        // Initialize bullet palletes list
-        if (selectionFilter.Editor is not null) {
-            foreach (var bullets in selectionFilter.Editor.Fumen.Bullets
-                         .Concat<IBulletPalleteReferencable>(selectionFilter.Editor.Fumen.Bells)
-                         .Where(b => b.ReferenceBulletPallete is not null)
-                         .GroupBy(b => b.ReferenceBulletPallete)) {
-                FilterBulletPalettesItems.First(b => b.BulletPalette == bullets.Key).IncludedBullets.AddRange(bullets);
-            }
-        }
-    }
-
-    public bool IsMatch(ISelectableObject obj)
-    {
-        if (IsCriticalFilterMode != YesNoFilteringMode.Any && obj is ICriticalableObject criticalObj)
-            switch (criticalObj.IsCritical) {
-                case true when IsCriticalFilterMode == YesNoFilteringMode.No:
-                case false when IsCriticalFilterMode == YesNoFilteringMode.Yes:
-                    return false;
-            }
-
-
-        if (NodeTypeFilteringMode != LinePointTypeFilteringMode.Any && obj is ConnectableObjectBase connectableObj)
-            switch (connectableObj) {
-                case ConnectableChildObjectBase when NodeTypeFilteringMode == LinePointTypeFilteringMode.Head:
-                case ConnectableStartObject when NodeTypeFilteringMode == LinePointTypeFilteringMode.Tail:
-                    return false;
-            }
-
-
-        if (HoldPartFilteringMode != LinePointTypeFilteringMode.Any && obj is Hold or HoldEnd)
-            switch (obj) {
-                case HoldEnd when HoldPartFilteringMode == LinePointTypeFilteringMode.Head:
-                case Hold when HoldPartFilteringMode == LinePointTypeFilteringMode.Tail:
-                    return false;
-            }
-
-
-        if (TagFilter != string.Empty) {
-            var tag = ((OngekiObjectBase)obj).Tag;
-            if (IsTagFilterRegex) {
-                if (!Regex.IsMatch(tag, TagFilter))
-                    return false;
-            }
-            else if (tag == TagFilter) {
-                return false;
-            }
-        }
-
-        if (obj is ILaneDockable dockable && FilterDockableLaneOptions.Any(o => o.IsSelected && !o.Filter(dockable))) {
-            return false;
-        }
-
-        if (obj is Flick flick && FlickDirectionFilteringMode != LeftRightFilteringMode.Any) {
-            if (FlickDirectionFilteringMode.ToFlickDirection() != flick.Direction) {
-                return false;
-            }
-        }
-
-        if (obj is IBulletPalleteReferencable bplObj
-            && FilterBulletPalettesItems.Any(f => f.IsSelected)
-            && FilterBulletPalettesItems.Any(f => !f.IsSelected && bplObj.ReferenceBulletPallete == f.BulletPalette)) {
-            return false;
-        }
-
-        if (OnlyMatchLanesWithSelectedStartNodes && obj is ConnectableChildObjectBase laneNext) {
-            if (!SelectionFilter.Editor.SelectObjects.Contains(laneNext.ReferenceStartObject)) {
-                return false;
-            }
-        } else if (OnlyMatchLanesWithSelectedStartNodes && obj is HoldEnd holdEnd) {
-            if (!SelectionFilter.Editor.SelectObjects.Contains(holdEnd.ReferenceLaneStart)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-}
-
 public sealed class FilterObjectTypeCategory : PropertyChangedBase
 {
     private readonly SelectionFilterViewModel Filter;
@@ -452,7 +246,7 @@ public class FilterObjectTypesItem : PropertyChangedBase
 {
     private readonly SelectionFilterViewModel SelectionFilter;
 
-    public List<ISelectableObject> MatchingObjects { get; private set; } = new();
+    public List<ISelectableObject> MatchingObjects { get; } = new();
 
     public FilterObjectTypesItem(SelectionFilterViewModel selectionFilter)
     {
@@ -474,55 +268,30 @@ public class FilterObjectTypesItem : PropertyChangedBase
 
 public class FilterBulletPalettesItem : PropertyChangedBase
 {
-    public BulletPallete BulletPalette;
-    public ObservableCollection<IBulletPalleteReferencable> IncludedBullets = new();
+    public BulletPallete BulletPalette { get; }
 
-    public string Display => $"{BulletPalette.StrID} {(string.IsNullOrWhiteSpace(BulletPalette.EditorName) ? "" : BulletPalette.EditorName + " ")}({IncludedBullets.Count})";
+    public List<ISelectableObject> MatchingBullets { get; private set; } = new();
+    public List<ISelectableObject> MatchingBells { get; private set; } = new();
 
-    private bool _isSelected = false;
+    public string Display =>
+        (BulletPalette is null
+            ? "No palette"
+            : $"{BulletPalette.StrID}{(string.IsNullOrWhiteSpace(BulletPalette.EditorName) ? "" : $" {BulletPalette.EditorName}")}")
+        + $" ({MatchingBullets.Count} | {MatchingBells.Count})";
+
+    public int ItemsCount => MatchingBullets.Count + MatchingBells.Count;
 
     public FilterBulletPalettesItem(BulletPallete bulletPalette)
     {
         BulletPalette = bulletPalette;
     }
 
+    private bool _isSelected = false;
     public bool IsSelected
     {
         get => _isSelected;
         set => Set(ref _isSelected, value);
     }
-}
-
-public enum LinePointTypeFilteringMode
-{
-    Head,
-    Tail,
-    Any
-}
-
-public enum YesNoFilteringMode
-{
-    Yes,
-    No,
-    Any
-}
-
-public enum LeftRightFilteringMode
-{
-    Left,
-    Right,
-    Any
-}
-
-public static class FilteringEnumExtensions
-{
-    public static Flick.FlickDirection ToFlickDirection(this LeftRightFilteringMode @this)
-        => @this switch
-        {
-            LeftRightFilteringMode.Left => Flick.FlickDirection.Left,
-            LeftRightFilteringMode.Right => Flick.FlickDirection.Right,
-            _ => throw new ArgumentOutOfRangeException(nameof(@this), @this, null)
-        };
 }
 
 public class ParamValueToBoolConverter : IValueConverter
