+++ OngekiFumenEditor/OngekiFumenEditor.csproj
@@ -102,7 +102,7 @@
 		<Resource Include="Resources\Icons\Batch\LaneBlock.png" />
 		<None Remove="Resources\Icons\Batch\Clipboard.png" />
 		<Resource Include="Resources\Icons\Batch\Clipboard.png" />
-		<Page Update="Modules\FumenEditorSelectingObjectViewer\Views\SelectionFilterWindowView.xaml">
+		<Page Update="Modules\SelectionFilter\Views\SelectionFilterWindowView.xaml">
 		  <Generator>MSBuild:Compile</Generator>
 		  <XamlRuntime>Wpf</XamlRuntime>
 		  <SubType>Designer</SubType>
